+++ app/minime/models.py
@@ -2,10 +2,26 @@ from django.db import models
 
 
 class Url(models.Model):
-    short_id = models.SlugField(max_length=6, primary_key=True)
-    httpurl = models.URLField(max_length=2048)
-    pub_date = models.DateTimeField(auto_now=True)
-    count = models.IntegerField(default=0)
+    hash = models.SlugField(max_length=6)
+    url = models.URLField(max_length=2048)
+    created = models.DateTimeField(auto_now=True)
+    password = models.CharField(max_length=1024, null=True)
 
     def __str__(self):
-        return self.httpurl
+        return self.url
+
+    class Meta:
+        indexes = [
+            models.Index(fields=['hash'])
+        ]
+
+
+class Visitors(models.Model):
+    url = models.ForeignKey(Url, on_delete=models.CASCADE)
+    browser_family = models.CharField(max_length=100)
+    browser_version = models.CharField(max_length=100)
+    os_family = models.CharField(max_length=100)
+    os_version = models.CharField(max_length=100)
+    device = models.CharField(max_length=100)
+    platform = models.CharField(max_length=100)
+    is_mobile = models.BooleanField()