+++ app/minime/urls.py
@@ -2,9 +2,11 @@ from django.conf.urls import url
 
 from . import views
 
+
 app_name = 'minime'
+
 urlpatterns = [
     url(r'^$', views.index, name='home'),
-    url(r'^r/(?P<short_id>\w{6})$', views.redirect, name='redirect'),
-    url(r'^shorten/$', views.shorten_url, name='shorten'),
+    url(r'^r/(?P<hash>\w{6})$', views.redirect, name='redirect'),
+    url(r'^shorten/$', views.CreateShortUrl.as_view(), name='shorten'),
 ]