+++ HISTORY.md
@@ -3,6 +3,9 @@
 
 ### Next
 
+- Save visitor stats on redirect. 
+- Added password protection.
+- Added Django Rest Framework.
 - Added test coverage and coveralls.
 - Added redis cache.
 - Added ability to run app on Heroku.